+++ src/app/compare/CompareStat.tsx
@@ -6,6 +6,8 @@ import { AppIcon, starIcon } from '../shell/icons';
 import clsx from 'clsx';
 import { t } from 'app/i18next-t';
 import RecoilStat from 'app/item-popup/RecoilStat';
+import { energyCapacityTypeNames } from '../item-popup/EnergyMeter';
+import ElementIcon from 'app/inventory/ElementIcon';
 
 export default function CompareStat({
   stat,
@@ -28,6 +30,9 @@ export default function CompareStat({
     >
       <span>
         {stat.id === 'Rating' && <AppIcon icon={starIcon} />}
+        {stat.id === 'EnergyCapacity' && itemStat && (
+          <ElementIcon element={energyCapacityTypeNames[itemStat.statHash]} />
+        )}
         {itemStat && itemStat.value !== undefined ? (
           itemStat.statHash === 2715839340 ? (
             <span className="stat-recoil">