+++ src/app/loadout-builder/LoadoutBuilder.tsx
@@ -12,7 +12,7 @@ import { DimStore, D2Store } from '../inventory/store-types';
 import { RootState } from '../store/reducers';
 import GeneratedSets from './generated-sets/GeneratedSets';
 import { filterGeneratedSets, isLoadoutBuilderItem } from './generated-sets/utils';
-import { ArmorSet, StatTypes, MinMax, ItemsByBucket, LockedMap } from './types';
+import { ArmorSet, StatTypes, ItemsByBucket, LockedMap, MinMaxIgnored } from './types';
 import { sortedStoresSelector, storesLoadedSelector, storesSelector } from '../inventory/reducer';
 import { process, filterItems, statKeys } from './process';
 import { createSelector } from 'reselect';
@@ -58,7 +58,7 @@ interface State {
   requirePerks: boolean;
   lockedMap: LockedMap;
   selectedStoreId?: string;
-  statFilters: Readonly<{ [statType in StatTypes]: MinMax }>;
+  statFilters: Readonly<{ [statType in StatTypes]: MinMaxIgnored }>;
   minimumPower: number;
   query: string;
   statOrder: StatTypes[];
@@ -124,12 +124,12 @@ export class LoadoutBuilder extends React.Component<Props & UIViewInjectedProps,
       requirePerks: true,
       lockedMap: {},
       statFilters: {
-        Mobility: { min: 0, max: 10 },
-        Resilience: { min: 0, max: 10 },
-        Recovery: { min: 0, max: 10 },
-        Discipline: { min: 0, max: 10 },
-        Intellect: { min: 0, max: 10 },
-        Strength: { min: 0, max: 10 }
+        Mobility: { min: 0, max: 10, ignored: false },
+        Resilience: { min: 0, max: 10, ignored: false },
+        Recovery: { min: 0, max: 10, ignored: false },
+        Discipline: { min: 0, max: 10, ignored: false },
+        Intellect: { min: 0, max: 10, ignored: false },
+        Strength: { min: 0, max: 10, ignored: false }
       },
       minimumPower: 750,
       query: '',
@@ -198,6 +198,9 @@ export class LoadoutBuilder extends React.Component<Props & UIViewInjectedProps,
     let combos = 0;
     let combosWithoutCaps = 0;
     let processError;
+    const enabledStats = new Set(
+      statKeys.filter((statType) => !this.state.statFilters[statType].ignored)
+    );
     try {
       filteredItems = this.filterItemsMemoized(
         items[store.classType],
@@ -205,7 +208,7 @@ export class LoadoutBuilder extends React.Component<Props & UIViewInjectedProps,
         lockedMap,
         filter
       );
-      const result = this.processMemoized(filteredItems);
+      const result = this.processMemoized(filteredItems, store.id);
       processedSets = result.sets;
       combos = result.combos;
       combosWithoutCaps = result.combosWithoutCaps;
@@ -214,7 +217,8 @@ export class LoadoutBuilder extends React.Component<Props & UIViewInjectedProps,
         minimumPower,
         lockedMap,
         statFilters,
-        statOrder
+        statOrder,
+        enabledStats
       );
     } catch (e) {
       console.error(e);
@@ -293,6 +297,7 @@ export class LoadoutBuilder extends React.Component<Props & UIViewInjectedProps,
               onLockedMapChanged={this.onLockedMapChanged}
               defs={defs}
               statOrder={statOrder}
+              enabledStats={enabledStats}
             />
           )}
         </PageWithMenu.Contents>
@@ -319,12 +324,12 @@ export class LoadoutBuilder extends React.Component<Props & UIViewInjectedProps,
       lockedMap: {},
       requirePerks: true,
       statFilters: {
-        Mobility: { min: 0, max: 10 },
-        Resilience: { min: 0, max: 10 },
-        Recovery: { min: 0, max: 10 },
-        Discipline: { min: 0, max: 10 },
-        Intellect: { min: 0, max: 10 },
-        Strength: { min: 0, max: 10 }
+        Mobility: { min: 0, max: 10, ignored: false },
+        Resilience: { min: 0, max: 10, ignored: false },
+        Recovery: { min: 0, max: 10, ignored: false },
+        Discipline: { min: 0, max: 10, ignored: false },
+        Intellect: { min: 0, max: 10, ignored: false },
+        Strength: { min: 0, max: 10, ignored: false }
       },
       minimumPower: 0
     });